requirements.txt
streamlit>=1.24.0
openai>=0.27.8
PyPDF2==3.0.1
python-dotenv==1.0.0
flask>=2.0.1
pandas>=2.0.0
numpy>=1.24.3